+++ changelog.md
@@ -2,7 +2,7 @@
 - Fixed the popup to fit based on your screen resolution
 - Removed **Edit Camera** due to issues with crashes on levels and compliances with the Geode Index guidelines. sorry :(
 - Fixed TextBox inputs not allowing user to set an decimal point on certain actions
-- Added **Set Gravity** & **Speed Player** Events
+- Added **Set Gravity**,**Speed Player** & **Noclip** Events
 # v0.1.5-beta.1
 - Added **Sound Effect** & **Stop All Sounds** Events
 - Changed the **Wait** Action value to use float value instead of just integer.